simplify GroupedCallReducer
