+++ inst/include/dplyr/Result/all.h
@@ -16,7 +16,6 @@
 
 #include <dplyr/Result/CallElementProxy.h>
 #include <dplyr/Result/DelayedProcessor.h>
-#include <dplyr/Result/CallbackProcessor.h>
 #include <dplyr/Result/ILazySubsets.h>
 #include <dplyr/Result/LazySubsets.h>
 #include <dplyr/Result/LazyGroupedSubsets.h>
